requirements.txt
annotated-types
anyio
Brotli
certifi
cffi
charset-normalizer
click
colorama
cryptography
dnspython
ecdsa
email-validator
gmpy2
greenlet
h11
h2
hpack
httpcore
httptools
httpx
hyperframe
idna
Jinja2
markdown-it-py
MarkupSafe
mdurl
pyasn1
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
Pygments
PySocks
python-dotenv
python-jose
python-multipart
PyYAML
requests
rich
rich-toolkit
rsa
shellingham
six
sniffio
SQLAlchemy
starlette
tomli
typer==0.21.1
typer-slim==0.21.1
typing_extensions
urllib3
watchfiles
websockets
win_inet_pton
zstandard==0.23.0
uvicorn[standard]
gunicorn
fastapi
fastapi-cli
azure-identity
azure-keyvault-secrets  # Add this if you are using Key Vault
finnhub-python  
azure-ai-agents 
azure-ai-projects 
azure-ai-inference